moving parsed_resources_tests file + some re-orgs
